upgraded Domatia to V2.0.0

+++ database/migrations/2025_04_11_094626_make_properties_fields_nullable.php
@@ -32,18 +32,9 @@ return new class extends Migration
 public function down()
 {
     Schema::table('properties', function (Blueprint $table) {
-        $table->string('location')->nullable(false)->change();
-        $table->decimal('price', 10, 2)->nullable(false)->change();
-        $table->string('tipo')->nullable(false)->change();
-        $table->text('description')->nullable(false)->change();
-        $table->text('description_en')->nullable(false)->change();
-        $table->text('description_fr')->nullable(false)->change();
-        $table->text('description_de')->nullable(false)->change();
-        $table->text('description_ru')->nullable(false)->change();
-        $table->dropColumn('thumbnail');
-        $table->string('ref')->nullable(false)->change();
-        $table->foreignId('zona_id')->nullable(false)->change();
-        $table->foreignId('propietario_id')->nullable(false)->change();
+        if (Schema::hasColumn('properties', 'thumbnail')) {
+            $table->dropColumn('thumbnail');
+        }
     });
 }
